requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==5.1.7
django-ckeditor==6.7.3
django-js-asset==3.1.2
emoji==2.11.0
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2024.2.0
future==0.18.3
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
mysql-connector==2.2.9
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
paytm-pg==1.1.2
pillow==10.2.0
prettytable==3.9.0
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.23.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.2
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.9
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
WTForms==3.1.2
